refactor: move item Hook onUpdate to method _onUpdate
